refactor(assignment_1): drop use of static_cast

+++ assignments/assignment_1/src/two_sum.cc
@@ -20,10 +20,10 @@ std::vector<int> two_sum(const std::vector<int> &nums, int target) {
 }
 
 std::vector<int> two_sum_naive(const std::vector<int> &nums, int target) {
-  for (size_t i = 0; i < nums.size(); i++) {
-    for (size_t j = 0; j < nums.size(); j++) {
+  for (int i = 0; i < nums.size(); i++) {
+    for (int j = 0; j < nums.size(); j++) {
       if (nums[i] + nums[j] == target)
-        return {static_cast<int>(i), static_cast<int>(j)};
+        return {i, j};
     }
   }
   return {};
